+++ modules/sql-engine/src/testFixtures/java/org/apache/ignite/internal/sql/engine/util/tpch/TpchHelper.java
@@ -65,32 +65,6 @@ public final class TpchHelper {
         }
     }
 
-    /**
-     * Loads query plan for provided TPC-H query id.
-     *
-     * @see #getQuery(String) for query id details.
-     */
-    public static String getQueryPlan(String queryId) {
-        // variant query ends with "v"
-        boolean variant = queryId.endsWith("v");
-        int numericId;
-
-        if (variant) {
-            String idString = queryId.substring(0, queryId.length() - 1);
-            numericId = Integer.parseInt(idString);
-        } else {
-            numericId = Integer.parseInt(queryId);
-        }
-
-        if (variant) {
-            var variantQueryFile = String.format("tpch/plan/variant_q%d.plan", numericId);
-            return loadFromResource(variantQueryFile);
-        } else {
-            var queryFile = String.format("tpch/plan/q%s.plan", numericId);
-            return loadFromResource(queryFile);
-        }
-    }
-
     static String loadFromResource(String resource) {
         try (InputStream is = TpchHelper.class.getClassLoader().getResourceAsStream(resource)) {
             if (is == null) {
